+++ src/main/kotlin/com/github/simiacryptus/aicoder/actions/code/InsertImplementationAction.kt
@@ -99,7 +99,7 @@ class InsertImplementationAction : BaseAction() {
     )
 
     override fun isEnabled(event: AnActionEvent): Boolean {
-        if (APIClientBase.isSanctioned()) return false
+        if (UITools.isSanctioned()) return false
         val computerLanguage = ComputerLanguage.getComputerLanguage(event) ?: return false
         if (computerLanguage == ComputerLanguage.Text) return false
         return getPsiClassContextActionParams(event).isPresent