+++ src/main/kotlin/com/github/simiacryptus/aicoder/actions/code/QuestionAction.kt
@@ -38,7 +38,7 @@ class QuestionAction : BaseAction() {
         ).create()
 
     override fun isEnabled(event: AnActionEvent): Boolean {
-        if (APIClientBase.isSanctioned()) return false
+        if (UITools.isSanctioned()) return false
         val computerLanguage = ComputerLanguage.getComputerLanguage(event) ?: return false
         return computerLanguage != ComputerLanguage.Text
     }