+++ src/main/kotlin/com/github/simiacryptus/aicoder/config/AppSettingsComponent.kt
@@ -79,8 +79,13 @@ class AppSettingsComponent {
     @Name("API Base")
     val apiBase = JBTextField()
 
+    @Name("File Actions")
+    var fileActions = ActionTable(AppSettingsState.instance.fileActions.actionSettings.values.map { it.copy() }
+        .toTypedArray().toMutableList())
+
     @Name("Editor Actions")
-    var editorActions = ActionTable()
+    var editorActions = ActionTable(AppSettingsState.instance.editorActions.actionSettings.values.map { it.copy() }
+        .toTypedArray().toMutableList())
 
     init {
         tokenCounter.isEditable = false