+++ src/main/kotlin/com/github/simiacryptus/aicoder/util/ComputerLanguage.kt
@@ -457,6 +457,7 @@ enum class ComputerLanguage(configuration: Configuration) {
     }
 
     companion object {
+        @JvmStatic
         fun findByExtension(extension: CharSequence): ComputerLanguage? {
             return Arrays.stream(values()).filter { x: ComputerLanguage ->
                 x.extensions.contains(
@@ -465,6 +466,7 @@ enum class ComputerLanguage(configuration: Configuration) {
             }.findAny().orElse(null)
         }
 
+        @JvmStatic
         fun getComputerLanguage(e: AnActionEvent?): ComputerLanguage? {
             val file = e?.getData(CommonDataKeys.VIRTUAL_FILE) ?: return null
             val extension = if (file.extension != null) file.extension!!.lowercase(Locale.getDefault()) else ""