+++ src/test/kotlin/com/github/simiacryptus/aicoder/actions/code/DocActionTest.kt
@@ -10,7 +10,7 @@ class DocActionTest : ActionTestBase() {
 
     @Test
     fun testProcessing() {
-        testActionScript(DocAction(), "/DocActionTest.md")
+        testScript_SelectionAction(DocAction(), "/DocActionTest.md")
     }
 
     @Test
@@ -23,7 +23,7 @@ class DocActionTest : ActionTestBase() {
     @Test
     fun testEditSelection() {
         val docAction = DocAction()
-        val editorState = SelectionAction.EditorState("fun hello() {\nprintln(\"Hello, world!\")\n}", 0, 10, 0, null)
+        val editorState = SelectionAction.EditorState("fun hello() {\nprintln(\"Hello, world!\")\n}", 0, Pair(0, 10), null, arrayOf())
         val result = docAction.editSelection(editorState, 0, 10)
         assertEquals(Pair(0, 10), result)
     }